refactoring / 各クラスの場所（名前空間）を分割・整理した。
