+++ MusicerBeat/Models/Services/SoundPlayerFactory.cs
@@ -1,4 +1,6 @@
-namespace MusicerBeat.Models
+using MusicerBeat.Models.Interfaces;
+
+namespace MusicerBeat.Models.Services
 {
     public class SoundPlayerFactory : ISoundPlayerFactory
     {
