+++ MusicerBeatTests/Models/VolumeControllerTest.cs
@@ -1,4 +1,5 @@
 using MusicerBeat.Models;
+using MusicerBeat.Models.Services;
 using MusicerBeatTests.ViewModels;
 
 namespace MusicerBeatTests.Models
