+++ MusicerBeatTests/ViewModels/DummySoundPlayerFactory.cs
@@ -1,4 +1,5 @@
 using MusicerBeat.Models;
+using MusicerBeat.Models.Interfaces;
 
 namespace MusicerBeatTests.ViewModels
 {
